+++ src/main.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './app/app';
-import '@/shared/dayjs';
+import '@/shared/helpers/dayJS';
 
 import '@/app/styles/index.scss';
